ciaa: Add openocd scripts to flash board
In order to allow the user to program the board's flash with all the
firmware in a simple fashion, we need a proper openocd configuration
file which initializes the internal flash and spifi devices.

This commit adds the CIAA NXP LPC 4337 openocd configuration and also
a user script to ease the firmware install task.

Signed-off-by: Martin Ribelotta <martinribelotta@gmail.com>
Signed-off-by: Ezequiel Garcia <ezequiel@vanguardiasur.com.ar>